requirements.txt
streamlit
google-genai
docxtpl
docx2pdf
mammoth
pdfplumber
requests
pytz
holidays
typing_extensions